sample_test_of_means
To perform One Sample Hypothesis Testing using Parametric & Non Parametric methods with the help of Scipy package. The following steps are to be followed for solving a On Sample Test of Means problem.

1. Formulate the Null and Alternate Hypothesis
For example, if you are working at a Piston Manufacturing plant and would like to test the quality of manufactured Pistons by observing the average diameter size of pistons is 20cm. The Null hypothesis would be

H0 : Population Mean of piston's diameter = 20 cm

and the Alternate Hypothesis would be

H1 : Population Mean of piston's diameter is NOT EQUAL TO 20 cm

2. Collect the sample data for the above Hypothesis and verify if the sample follows a Normal Distribution.
If the Data follows a Normal Distribution, then we could solve the problem using a Parametric One Sample test such as the Student's t-test. Else, we would require the Wilcoxon Test which is better suited for Non Normal Distributions of data.

3. Run the appropriate test (Parametric vs Non Parametric) and observe the p-value of the same.
If p-value >= significance level (generally set at 0.05) , then we would fail to reject the H0

Alternatively, if p-value < significance level (generally set at 0.05) , then we could reject the H0 and prove the H1